Function mpcnet::command::command_handler

pub async fn command_handler(eventloop: &mut EventLoop, command: Command)

Handles incoming commands for the network event loop.

This async function processes various network-related commands and performs corresponding actions in the libp2p Swarm. It deals with network operations like listening, dialing, responding to requests, etc.

Arguments

eventloop - The mutable reference to the EventLoop managing the network operations.

command - The Command to be processed.

Examples

Handling a command in the network event loop:

command_handler(&mut eventloop, command).await;