Function mpcnet::sss::refresh_shares

pub fn refresh_shares(
    shares_map: &mut HashMap<u8, Vec<u8>>,
    threshold: usize
) -> Result<(), String>

https://en.wikipedia.org/wiki/Proactive_secret_sharing#Mathematics Refreshes the shares of a secret in a proactive secret sharing scheme.

This method allows the shares to be updated without changing the underlying secret, enhancing security against attacks that compromise share over time.

Arguments

shares_map - A mutable reference to a HashMap representing the shares.

threshold - The minimum number of shares required to reconstruct the secret.

Returns

Result<(), String> indicating successful completion or an error message.

Errors

Returns Err if threshold is less than or equal to 1.

Returns Err if the shares_map is empty.

Examples

Refreshing shares of a secret:

refresh_shares(&mut shares_map, 3).unwrap();
// The shares in `shares_map` are now refreshed.